pyproject.toml
[project]
name = "bams-ai-data-exploration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bokeh==3.3.0",
    "datashader>=0.18.2",
    "holoviews>=1.22.1",
    "ipykernel>=7.2.0",
    "lancedb>=0.29.2",
    "matplotlib>=3.10.8",
    "notebook>=7.5.4",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "panel==1.3.1",
    "pillow>=12.1.1",
    "pyarrow>=23.0.1",
    "pylance>=2.0.1",
    "pyviz-comms==3.0.0",
    "open-clip-torch>=2.24.0",
    "timm>=1.0.25",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "marimo>=0.13.0",
    "rich>=13.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F"]
ignore = [
    "E501",  # line too long — not enforced
    "E402",  # module-level import not at top (notebooks use late imports)
    "E731",  # lambda assignment
    "E741",  # ambiguous variable name (i, l, O etc.)
]
